+++ src/lib/generate.ts
@@ -22,12 +22,7 @@ export const generateBBAN = (
   accountNumber: ProbablyString,
   blz: ProbablyString,
 ): string | null => {
-  if (
-    !accountNumber ||
-    !accountNumber.match(/^\d{1,10}$/) ||
-    !blz ||
-    !blz.match(/^[1-9]\d{7}$/)
-  ) {
+  if (!accountNumber?.match(/^\d{1,10}$/) || !blz?.match(/^[1-9]\d{7}$/)) {
     return null;
   }